make files smaller
